# Conflicts:
#	app/src/main/java/com/example/chillinapp/ChillInActivity.kt
#	app/src/main/java/com/example/chillinapp/data/stress/FirebaseStressDataDao.kt
#	app/src/main/java/com/example/chillinapp/data/stress/StressData.kt
#	build.gradle.kts

+++ app/src/main/java/com/example/chillinapp/data/stress/FirebaseStressDataDao.kt
@@ -47,7 +47,6 @@ class FirebaseStressDataDao {
 
             }
 
-
             val fastreturn=fastInsert(stressData)
             if (fastreturn.success)
             ServiceResult(true,null,null)
@@ -101,7 +100,7 @@ class FirebaseStressDataDao {
             querySnapshot?.forEach { document ->
                 val timestamp = document.id.toLongOrNull()
                 if (timestamp != null) {
-                    val heartrateSensor:Double = (document.get("heartrateSensor") as Double)
+                    val heartrateSensor:Float = (document.get("heartrateSensor") as Float)
                     val skinTemperatureSensor : Double= (document.get("skinTemperatureSensor" ) as Double)
 
                     // Costruisci l'oggetto StressRawData e aggiungilo alla lista
@@ -110,7 +109,6 @@ class FirebaseStressDataDao {
                 }
 
             }
-
 
             ServiceResult(true, rawDataList, null)
         } catch (e: Exception) {
@@ -170,11 +168,7 @@ class FirebaseStressDataDao {
             }
         }
        else
-
-        return ServiceResult(success = false, data = null, error = StressErrorType.NOACCOUNT)
-
-
-
+           return ServiceResult(success = false, data = null, error = StressErrorType.NOACCOUNT)
     }
     suspend fun fastget():ServiceResult<List<StressRawData>,StressErrorType>{
         val user=auth.currentUser
@@ -186,7 +180,7 @@ class FirebaseStressDataDao {
             val stressDataList= mutableListOf<StressRawData>()
             snapshot?.children?.forEach{ childSnapshot->
                 val timestamp=childSnapshot.key?.toLongOrNull()
-                val heartrateSensor=childSnapshot.child("heartrateSensor").getValue() as Double
+                val heartrateSensor=childSnapshot.child("heartrateSensor").getValue() as Float
                 val skinTemperatureSensor=childSnapshot.child("skinTemperatureSensor").getValue() as Double
                 if (timestamp != null && heartrateSensor != null && skinTemperatureSensor != null) {
                     val stressData = StressRawData(timestamp, heartrateSensor, skinTemperatureSensor)
@@ -196,11 +190,10 @@ class FirebaseStressDataDao {
 
             ServiceResult(success = true, data = stressDataList, error = null)
 
-            }
-        catch (e: Exception) {
+        } catch (e: Exception) {
             Log.e("readStressDataFromFirebase", "An exception occurred", e)
             ServiceResult(success = false, data = null, error = StressErrorType.COMMUNICATION_PROBLEM)
         }
 
     }
-    }
+}
